Rewritten SOAP example
